CONFETI-43 feat: 온보딩 완료 API 엔드포인트 변경 - 기존 auth/v4/onboard 에서 user/onboard/v4 로 변경. 다른 온보딩 API 등과 통일
