+++ src/main/java/org/sopt/confeti/api/user/facade/UserOnboardFacade.java
@@ -15,6 +15,8 @@ import org.sopt.confeti.api.user.facade.dto.response.onboard.GetOnboardStatusDTO
 import org.sopt.confeti.api.user.facade.dto.response.onboard.UserOnboardCacheDTO;
 import org.sopt.confeti.api.user.facade.dto.response.onboard.UserOnboardFavoriteArtistsDTO;
 import org.sopt.confeti.api.user.facade.dto.response.onboard.UserOnboardRelatedArtistsDTO;
+import org.sopt.confeti.domain.artist_favorite.application.ArtistFavoriteService;
+import org.sopt.confeti.domain.user.User;
 import org.sopt.confeti.domain.user.application.UserOnboardService;
 import org.sopt.confeti.domain.user.application.UserService;
 import org.sopt.confeti.domain.user.constant.Role;
@@ -44,6 +46,7 @@ public class UserOnboardFacade {
     private final UserOnboardService userOnboardService;
     private final RedisTemplate<String, Object> redisTemplate;
     private final ObjectMapper objectMapper;
+    private final ArtistFavoriteService artistFavoriteService;
 
     public UserOnboardRelatedArtistsDTO getArtistsRelatedTerm(long userId, String term, int limit) {
         Set<String> topArtistIds = userOnboardService.getCachedExposedArtistIds(userId);
@@ -147,6 +150,20 @@ public class UserOnboardFacade {
             cachedArtists.favoriteArtistIds());
 
         return UserOnboardFavoriteArtistsDTO.from(favoriteArtists);
+    }
+
+    @Transactional
+    public void onboard(long userId) {
+        UserOnboardCacheDTO cachedArtists = userOnboardService.getCachedArtists(userId);
+        Set<String> favoriteArtistIds = cachedArtists.favoriteArtistIds();
+        User user = userService.findById(userId);
+
+        artistFavoriteService.addFavorites(user, favoriteArtistIds);
+        user.setRole(Role.GENERAL);
+    }
+
+    public void flushCachedOnboardArtists(long userId) {
+        userOnboardService.flushCachedOnboardArtists(userId);
     }
 
     public void cacheTopArtists(List<ConfetiArtist> topArtists) {
